+++ src/main/java/com/doublew2w/sbs/mybatis/binding/MapperProxy.java
@@ -5,6 +5,7 @@ import java.io.Serializable;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.util.Map;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * 映射器代理类
@@ -13,6 +14,7 @@ import java.util.Map;
  * @date: 2024/9/1 4:33
  * @project: sbs-mybatis
  */
+@Slf4j
 public class MapperProxy<T> implements InvocationHandler, Serializable {
 
   private static final long serialVersionUID = -8932997989399317814L;
@@ -40,6 +42,10 @@ public class MapperProxy<T> implements InvocationHandler, Serializable {
       return method.invoke(this, args);
     } else {
       final MapperMethod mapperMethod = cachedMapperMethod(method);
+      log.info(
+          "into invoke(), mapperInterface: {}, cachedMapperMethod: {}",
+          mapperInterface.getName(),
+          method.getName());
       return mapperMethod.execute(sqlSession, args);
     }
   }