+++ src/test/java/com/doublew2w/sbs/mybatis/test/IUserDaoApiTest.java
@@ -45,4 +45,21 @@ public class IUserDaoApiTest {
     User user = userDao.queryUserInfoById(1L);
     logger.info("测试结果：{}", JSON.toJSONString(user));
   }
+
+  @Test
+  public void test_PooledDataSource() throws Exception {
+    // 1. 从SqlSessionFactory中获取SqlSession
+    SqlSessionFactory sqlSessionFactory =
+        new SqlSessionFactoryBuilder().build(Resources.getResourceAsReader("mybatis-config.xml"));
+    SqlSession sqlSession = sqlSessionFactory.openSession();
+
+    // 2. 获取映射器对象
+    IUserDao userDao = sqlSession.getMapper(IUserDao.class);
+
+    // 3. 测试验证
+    for (int i = 0; i < 10; i++) {
+      User user = userDao.queryUserInfoById(1L);
+      logger.info("测试结果：{}", JSON.toJSONString(user));
+    }
+  }
 }